Title: Modified errorHandler added languajes
Description:
Modified errorHandler library in order to support languajes. Now it receive
an Error object with the status number, the text key and the languaje.

Modified controller for supporting this new error handling

Added two files with the dictionary for english and spanish

Added a middleware in app.js for getting the languaje in the request

 - modified:   app.js
 - new file:   lib/errorDictionaryEng.js
 - new file:   lib/errorDictionaryEsp.js
 - modified:   lib/errorHandler.js
 - modified:   routes/api/v1/commercials.js
 - modified:   routes/api/v1/tokens.js
 - modified:   routes/api/v1/users.js

Reported-by: -
Signed-off-by: Juan Arillo <juanarillo@outlook.com>